fix(jsx/dom): Intrinsic elements can now also be safely Evaluated multiple times for the same element
