+++ src/jsx/dom/jsx-dev-runtime.ts
@@ -1,31 +1,16 @@
 import type { Props, JSXNode } from '../base'
 import { normalizeIntrinsicElementProps } from '../utils'
-
-const JSXNodeCompatPrototype = {
-  type: {
-    get(this: { tag: string | Function }): string | Function {
-      return this.tag
-    },
-  },
-  ref: {
-    get(this: { props?: { ref: unknown } }): unknown {
-      return this.props?.ref
-    },
-  },
-}
+import { newJSXNode } from './utils'
 
 export const jsxDEV = (tag: string | Function, props: Props, key?: string): JSXNode => {
   if (typeof tag === 'string') {
     normalizeIntrinsicElementProps(props)
   }
-  return Object.defineProperties(
-    {
-      tag,
-      props,
-      key,
-    },
-    JSXNodeCompatPrototype
-  ) as JSXNode
+  return newJSXNode({
+    tag,
+    props,
+    key,
+  })
 }
 
 export const Fragment = (props: Record<string, unknown>): JSXNode => jsxDEV('', props, undefined)
